Make development tools code cross-platform.
