+++ Library/Homebrew/extend/os/development_tools.rb
@@ -1,0 +1,5 @@
+require "development_tools"
+
+if OS.mac?
+  require "extend/os/mac/development_tools"
+end
