+++ Library/Homebrew/os/mac/xcode.rb
@@ -120,7 +120,7 @@ module OS
           # https://github.com/Homebrew/brew/blob/master/share/doc/homebrew/Xcode.md
           "4.0"
         else
-          case (MacOS.clang_version.to_f * 10).to_i
+          case (DevelopmentTools.clang_version.to_f * 10).to_i
           when 0       then "dunno"
           when 1..14   then "3.2.2"
           when 15      then "3.2.4"
